backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-cloud-firestore==2.13.1
gunicorn==21.2.0
pytest==7.4.1
pytest-flask==1.2.0